backend/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain-google-genai
pinecone-client
pydantic
